code: clean up javac -Xlint warnings (unchecked, serial, redundant cast)
- TournamentAnalyzer.slaterExact: replace the generic-array sentinel

  `List<String>[] bestOrder = new List[]{...}` (which forced an

  `[unchecked] unchecked conversion` warning and the only suppression-

  free generic-array allocation in the file) with a type-safe

  single-element `List<List<String>>`. Same algorithm, same mutation

  semantics through the branch-and-bound recursion, zero warnings.

- GraphInfluenceSeedAdvisor.synthesize: drop two `(double)` casts on

  values that are already declared `double` (`expectedCoverage` and

  `baseline`). Flagged by `-Xlint:cast` as redundant.

- Add `serialVersionUID = 1L` to four Swing classes that extend

  `JComponent`/`JFrame`/`JDialog` (AdjacencyMatrixHeatmap, Main,

  StatsPanel, RandomGraphDialog). These are `Serializable` by

  inheritance and were each emitting `[serial] no definition of

  serialVersionUID`; the JDK uses an implementation-specific hash

  when the field is missing, so any future field reorder silently

  breaks deserialization.

Verification: `mvn clean compile` is clean of the targeted warnings

(unchecked / cast / serial). `mvn test` for the touched classes:

TournamentAnalyzerTest 76/76 pass, GraphInfluenceSeedAdvisorTest 14/14

pass. Full `mvn test` shows the same 27 failures + 8 errors as on

origin/master (LinkPredictionAnalyzerTest, GraphLabelingAnalyzerTest,

etc.) — pre-existing, unrelated to this change.

+++ Gvisual/src/gvisual/GraphInfluenceSeedAdvisor.java
@@ -428,12 +428,12 @@ public final class GraphInfluenceSeedAdvisor<V, E> {
         double reductionFraction = 0.0;
         if (mode == Mode.SPREAD) {
             expectedCoverage = finalDecisions.isEmpty() ? 0.0 : finalDecisions.get(finalDecisions.size() - 1).coverageEstimate;
-            coverageFraction = (double) expectedCoverage / n;
+            coverageFraction = expectedCoverage / n;
         } else {
             double remaining = finalDecisions.isEmpty() ? baseline : finalDecisions.get(finalDecisions.size() - 1).coverageEstimate;
             expectedCoverage = remaining;
             expectedReduction = baseline - remaining;
-            coverageFraction = (double) baseline / n;
+            coverageFraction = baseline / n;
             reductionFraction = baseline > 0 ? expectedReduction / baseline : 0.0;
         }
